Make UniqueFontID globally available
UniqueFontID will be useful to DrawObj.*, so pull it out of
CPlayBoard's internals.
